[Refactor] Split Java UDF runtime into focused modules
