+++ be/src/exprs/agg/java_udaf_function.h
@@ -33,7 +33,7 @@
 #include "jni.h"
 #include "types/logical_type.h"
 #include "udf/java/java_data_converter.h"
-#include "udf/java/java_udf.h"
+#include "udf/java/java_udaf_context.h"
 
 namespace starrocks {
