+++ be/src/exprs/java_function_call_expr.cpp
@@ -39,7 +39,9 @@
 #include "platform/user_function_cache.h"
 #include "types/type_descriptor.h"
 #include "udf/java/java_data_converter.h"
-#include "udf/java/java_udf.h"
+#include "udf/java/java_runtime.h"
+#include "udf/java/java_udf_context.h"
+#include "udf/java/jvm_function_helper.h"
 #include "udf/java/utils.h"
 
 namespace starrocks {
