+++ be/src/fs/hdfs/hdfs_fs_cache.cpp
@@ -20,7 +20,7 @@
 #include "fs/fs_options_helper.h"
 #include "fs/hdfs/hdfs_util.h"
 #include "gutil/strings/substitute.h"
-#include "udf/java/java_udf.h"
+#include "udf/java/java_runtime.h"
 
 namespace starrocks {
