+++ be/src/udf/java/java_data_converter.cpp
@@ -39,7 +39,7 @@
 #include "types/logical_type.h"
 #include "types/timestamp_value.h"
 #include "types/type_descriptor.h"
-#include "udf/java/java_udf.h"
+#include "udf/java/jvm_function_helper.h"
 #include "udf/java/type_traits.h"
 
 #define APPLY_FOR_NUMBERIC_TYPE(M) \
