+++ be/src/udf/java/jvm_function_helper.cpp
@@ -12,24 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "udf/java/java_udf.h"
+#include "udf/java/jvm_function_helper.h"
 
+#include <algorithm>
 #include <iterator>
 #include <memory>
-#include <sstream>
 #include <string>
 
-#include "base/utility/defer_op.h"
-#include "column/binary_column.h"
 #include "column/column.h"
 #include "common/status.h"
-#include "exprs/function_context.h"
 #include "fmt/core.h"
 #include "jni.h"
 #include "types/date_value.h"
 #include "types/logical_type.h"
 #include "types/timestamp_value.h"
+#include "udf/java/java_call_stub.h"
 #include "udf/java/java_native_method.h"
+#include "udf/java/java_udaf_context.h"
 #include "udf/java/type_traits.h"
 #include "udf/java/utils.h"
 
@@ -120,49 +119,6 @@ static JNINativeMethod java_native_methods[] = {
 };
 #pragma GCC diagnostic pop
 
-JavaUDAFUniqueContext* get_java_udaf_context(FunctionContext* ctx) {
-    if (ctx == nullptr) {
-        return nullptr;
-    }
-    return reinterpret_cast<JavaUDAFUniqueContext*>(ctx->get_function_state(FunctionContext::THREAD_LOCAL));
-}
-
-void attach_java_udaf_context(FunctionContext* ctx, std::unique_ptr<JavaUDAFUniqueContext> udaf_ctx) {
-    DCHECK(ctx != nullptr);
-    DCHECK(udaf_ctx != nullptr);
-    auto* old_ctx = get_java_udaf_context(ctx);
-    DCHECK(old_ctx == nullptr) << "duplicate Java UDAF context attach";
-    if (old_ctx != nullptr) {
-        delete old_ctx;
-    }
-    ctx->set_function_state(FunctionContext::THREAD_LOCAL, udaf_ctx.release());
-}
-
-void clear_java_udaf_states(FunctionContext* ctx) {
-    auto* udaf_ctx = get_java_udaf_context(ctx);
-    if (udaf_ctx == nullptr || udaf_ctx->states == nullptr) {
-        return;
-    }
-
-    auto env = JVMFunctionHelper::getInstance().getEnv();
-    udaf_ctx->states->clear(ctx, env);
-}
-
-void destroy_java_udaf_context(FunctionContext* ctx) {
-    if (ctx == nullptr) {
-        return;
-    }
-
-    auto* udaf_ctx = get_java_udaf_context(ctx);
-    if (udaf_ctx == nullptr) {
-        return;
-    }
-
-    clear_java_udaf_states(ctx);
-    ctx->set_function_state(FunctionContext::THREAD_LOCAL, nullptr);
-    delete udaf_ctx;
-}
-
 StatusOr<jobject> MapMeta::newLocalInstance(jobject keys, jobject values) const {
     JNIEnv* env = getJNIEnv();
     auto res = env->NewObject(immutable_map_class->clazz(), immutable_map_constructor, keys, values);
@@ -177,6 +133,10 @@ JVMFunctionHelper& JVMFunctionHelper::getInstance() {
     }
     static JVMFunctionHelper helper;
     return helper;
+}
+
+std::string dump_jni_exception_string(jthrowable throwable) {
+    return JVMFunctionHelper::getInstance().dumpExceptionString(throwable);
 }
 
 std::pair<JNIEnv*, JVMFunctionHelper&> JVMFunctionHelper::getInstanceWithEnv() {
@@ -776,509 +736,6 @@ DirectByteBuffer::~DirectByteBuffer() {
     }
 }
 
-UDAFStateList::UDAFStateList(JavaGlobalRef&& handle, JavaGlobalRef&& get, JavaGlobalRef&& batch_get,
-                             JavaGlobalRef&& add, JavaGlobalRef&& remove, JavaGlobalRef&& clear)
-        : _handle(std::move(handle)),
-          _get_method(std::move(get)),
-          _batch_get_method(std::move(batch_get)),
-          _add_method(std::move(add)),
-          _remove_method(std::move(remove)),
-          _clear_method(std::move(clear)) {
-    auto* env = JVMFunctionHelper::getInstance().getEnv();
-    _get_method_id = env->FromReflectedMethod(_get_method.handle());
-    _batch_get_method_id = env->FromReflectedMethod(_batch_get_method.handle());
-    _add_method_id = env->FromReflectedMethod(_add_method.handle());
-    _remove_method_id = env->FromReflectedMethod(_remove_method.handle());
-    _clear_method_id = env->FromReflectedMethod(_clear_method.handle());
-}
-
-jobject UDAFStateList::get_state(FunctionContext* ctx, JNIEnv* env, int state_handle) {
-    auto obj = env->CallObjectMethod(_handle.handle(), _get_method_id, state_handle);
-    CHECK_UDF_CALL_EXCEPTION(env, ctx);
-    return obj;
-}
-
-jobject UDAFStateList::get_state(FunctionContext* ctx, JNIEnv* env, jobject state_ids) {
-    auto obj = env->CallObjectMethod(_handle.handle(), _batch_get_method_id, state_ids);
-    CHECK_UDF_CALL_EXCEPTION(env, ctx);
-    return obj;
-}
-
-int UDAFStateList::add_state(FunctionContext* ctx, JNIEnv* env, jobject state) {
-    auto res = env->CallIntMethod(_handle.handle(), _add_method_id, state);
-    CHECK_UDF_CALL_EXCEPTION(env, ctx);
-    return res;
-}
-
-void UDAFStateList::remove(FunctionContext* ctx, JNIEnv* env, int state_handle) {
-    env->CallVoidMethod(_handle.handle(), _remove_method_id, state_handle);
-    CHECK_UDF_CALL_EXCEPTION(env, ctx);
-}
-
-void UDAFStateList::clear(FunctionContext* ctx, JNIEnv* env) {
-    env->CallVoidMethod(_handle.handle(), _clear_method_id);
-    CHECK_UDF_CALL_EXCEPTION(env, ctx);
-}
-
-ClassLoader::~ClassLoader() {
-    _handle.clear();
-    _clazz.clear();
-}
-
-constexpr const char* CLASS_LOADER_NAME = "com.starrocks.udf.UDFClassLoader";
-constexpr const char* CLASS_ANALYZER_NAME = "com.starrocks.udf.UDFClassAnalyzer";
-
-Status ClassLoader::init() {
-    JNIEnv* env = JVMFunctionHelper::getInstance().getEnv();
-    if (env == nullptr) {
-        return Status::InternalError("Init JNIEnv fail");
-    }
-    std::string name = JVMFunctionHelper::to_jni_class_name(CLASS_LOADER_NAME);
-    jclass clazz = env->FindClass(name.c_str());
-    LOCAL_REF_GUARD(clazz);
-
-    if (clazz == nullptr) {
-        RETURN_ERROR_IF_JNI_EXCEPTION_WITH_PREFIX(env, "ClassLoader Not Found");
-        return Status::InternalError(fmt::format("ClassLoader Not Found: {}", CLASS_LOADER_NAME));
-    }
-    _clazz = env->NewGlobalRef(clazz);
-
-    jmethodID udf_loader_contructor = env->GetMethodID(clazz, "<init>", "(Ljava/lang/String;)V");
-    if (udf_loader_contructor == nullptr) {
-        RETURN_ERROR_IF_JNI_EXCEPTION_WITH_PREFIX(env, "ClassLoader constructor Not Found");
-        return Status::InternalError("ClassLoader constructor Not Found");
-    }
-
-    // create class loader instance
-    jstring jstr = env->NewStringUTF(_path.c_str());
-    RETURN_ERROR_IF_JNI_EXCEPTION_WITH_PREFIX(env, "couldn't create jstring for classloader path");
-    LOCAL_REF_GUARD(jstr);
-
-    auto handle = env->NewObject(clazz, udf_loader_contructor, jstr);
-    RETURN_ERROR_IF_JNI_EXCEPTION_WITH_PREFIX(env, "couldn't create classloader");
-    LOCAL_REF_GUARD(handle);
-    _handle = env->NewGlobalRef(handle);
-
-    // init method id
-    _get_class = env->GetMethodID((jclass)_clazz.handle(), "findClass", "(Ljava/lang/String;)Ljava/lang/Class;");
-    if (_get_class == nullptr) {
-        RETURN_ERROR_IF_JNI_EXCEPTION_WITH_PREFIX(env, "couldn't get findClass method for classloader");
-        return Status::InternalError("couldn't get findClass method for classloader");
-    }
-
-    _get_call_stub =
-            env->GetMethodID((jclass)_clazz.handle(), "generateCallStubV",
-                             "(Ljava/lang/String;Ljava/lang/Class;Ljava/lang/reflect/Method;II)Ljava/lang/Class;");
-    if (_get_call_stub == nullptr) {
-        RETURN_ERROR_IF_JNI_EXCEPTION_WITH_PREFIX(env, "couldn't get generateCallStubV method for classloader");
-        return Status::InternalError("couldn't get generateCallStubV method for classloader");
-    }
-
-    return Status::OK();
-}
-
-StatusOr<JVMClass> ClassLoader::getClass(const std::string& className) {
-    auto& helper = JVMFunctionHelper::getInstance();
-    JNIEnv* env = helper.getEnv();
-    CHECK(env != nullptr) << "couldn't got a JNIEnv";
-
-    // class Name java.lang.Object -> java/lang/Object
-    std::string jni_class_name = JVMFunctionHelper::to_jni_class_name(className);
-    // invoke class loader
-    ASSIGN_OR_RETURN(jstring jstr_name, helper.to_jstring(jni_class_name));
-    LOCAL_REF_GUARD(jstr_name);
-
-    auto loaded_clazz = env->CallObjectMethod(_handle.handle(), _get_class, jstr_name);
-    LOCAL_REF_GUARD(loaded_clazz);
-
-    // check exception
-    RETURN_ERROR_IF_EXCEPTION(env, "exception happened when get class: {}");
-    // no exception happened, class exists
-    DCHECK(loaded_clazz != nullptr);
-
-    auto res = env->NewGlobalRef(loaded_clazz);
-    return res;
-}
-
-StatusOr<JVMClass> ClassLoader::genCallStub(const std::string& stubClassName, jclass clazz, jobject method, int type,
-                                            int numActualVarArgs) {
-    auto& helper = JVMFunctionHelper::getInstance();
-    JNIEnv* env = helper.getEnv();
-
-    std::string jni_class_name = JVMFunctionHelper::to_jni_class_name(stubClassName);
-    ASSIGN_OR_RETURN(jstring jstr_name, helper.to_jstring(jni_class_name));
-    LOCAL_REF_GUARD(jstr_name);
-
-    // generate call stub; pass numActualVarArgs for varargs methods
-    auto loaded_clazz =
-            env->CallObjectMethod(_handle.handle(), _get_call_stub, jstr_name, clazz, method, type, numActualVarArgs);
-    LOCAL_REF_GUARD(loaded_clazz);
-    RETURN_ERROR_IF_EXCEPTION(env, "exception happened when gen call stub: {}");
-
-    auto res = env->NewGlobalRef(loaded_clazz);
-    return res;
-}
-
-jmethodID JavaMethodDescriptor::get_method_id() const {
-    return JVMFunctionHelper::getInstance().getEnv()->FromReflectedMethod(method.handle());
-}
-
-Status ClassAnalyzer::has_method(jclass clazz, const std::string& method, bool* has) {
-    DCHECK(clazz != nullptr);
-    DCHECK(has != nullptr);
-
-    auto& helper = JVMFunctionHelper::getInstance();
-    JNIEnv* env = getJNIEnv();
-
-    std::string anlyzer_clazz_name = JVMFunctionHelper::to_jni_class_name(CLASS_ANALYZER_NAME);
-    jclass class_analyzer = env->FindClass(anlyzer_clazz_name.c_str());
-    LOCAL_REF_GUARD(class_analyzer);
-
-    if (class_analyzer == nullptr) {
-        return Status::InternalError(fmt::format("ClassAnalyzer Not Found: {}", CLASS_ANALYZER_NAME));
-    }
-
-    jmethodID hasMethod =
-            env->GetStaticMethodID(class_analyzer, "hasMemberMethod", "(Ljava/lang/String;Ljava/lang/Class;)Z");
-    if (hasMethod == nullptr) {
-        return Status::InternalError("couldn't found hasMethod method");
-    }
-
-    ASSIGN_OR_RETURN(jstring method_name, helper.to_jstring(method.c_str()));
-    LOCAL_REF_GUARD(method_name);
-
-    *has = env->CallStaticBooleanMethod(class_analyzer, hasMethod, method_name, (jobject)clazz);
-
-    RETURN_ERROR_IF_JNI_EXCEPTION_WITH_PREFIX(env, "call hasMemberMethod failed");
-
-    return Status::OK();
-}
-
-void ClassAnalyzer::strip_jni_generic_types(std::string* sign) {
-    std::string cleaned;
-    cleaned.reserve(sign->size());
-    int depth = 0;
-    for (char c : *sign) {
-        if (c == '<') {
-            depth++;
-        } else if (c == '>') {
-            depth--;
-        } else if (depth == 0) {
-            cleaned += c;
-        }
-    }
-    *sign = std::move(cleaned);
-}
-
-Status ClassAnalyzer::get_signature(jclass clazz, const std::string& method, std::string* sign) {
-    DCHECK(clazz != nullptr);
-    DCHECK(sign != nullptr);
-    auto& helper = JVMFunctionHelper::getInstance();
-    JNIEnv* env = helper.getEnv();
-    std::string anlyzer_clazz_name = JVMFunctionHelper::to_jni_class_name(CLASS_ANALYZER_NAME);
-    jclass class_analyzer = env->FindClass(anlyzer_clazz_name.c_str());
-    LOCAL_REF_GUARD(class_analyzer);
-
-    if (class_analyzer == nullptr) {
-        return Status::InternalError(fmt::format("ClassAnalyzer Not Found: {}", CLASS_ANALYZER_NAME));
-    }
-    jmethodID getSign = env->GetStaticMethodID(class_analyzer, "getSignature",
-                                               "(Ljava/lang/String;Ljava/lang/Class;)Ljava/lang/String;");
-    if (getSign == nullptr) {
-        return Status::InternalError("couldn't found getSignature method");
-    }
-
-    ASSIGN_OR_RETURN(jstring method_name, helper.to_jstring(method.c_str()));
-    LOCAL_REF_GUARD(method_name);
-
-    jobject result_sign = env->CallStaticObjectMethod(class_analyzer, getSign, method_name, (jobject)clazz);
-    LOCAL_REF_GUARD(result_sign);
-
-    RETURN_ERROR_IF_JNI_EXCEPTION_WITH_PREFIX(env, "call getSignature failed");
-
-    if (result_sign == nullptr) {
-        return Status::InternalError(fmt::format("couldn't found method:{}", method));
-    }
-    *sign = helper.to_string(result_sign);
-    // Strip generic type parameters from signature to produce standard JNI method descriptor.
-    // Java's getGenericParameterTypes() may produce signatures with generic info that:
-    // 1. JNI GetMethodID cannot match against the erased method descriptor, returning NULL.
-    // 2. get_udaf_method_desc() uses exact string matching (e.g. type == "java/util/List")
-    //    to populate method_desc. Generic signatures like "java/util/List<java/lang/String>"
-    //    would fail to match, causing missing MethodTypeDescriptor entries and subsequent
-    //    out-of-bounds access in process()/update()/merge() when indexing method_desc[j+1].
-    strip_jni_generic_types(sign);
-    return Status::OK();
-}
-
-StatusOr<jobject> ClassAnalyzer::get_method_object(jclass clazz, const std::string& method) {
-    auto& helper = JVMFunctionHelper::getInstance();
-    JNIEnv* env = helper.getEnv();
-
-    std::string anlyzer_clazz_name = JVMFunctionHelper::to_jni_class_name(CLASS_ANALYZER_NAME);
-    jclass class_analyzer = env->FindClass(anlyzer_clazz_name.c_str());
-    LOCAL_REF_GUARD(class_analyzer);
-    DCHECK(class_analyzer);
-
-    jmethodID getMethodObject = env->GetStaticMethodID(
-            class_analyzer, "getMethodObject", "(Ljava/lang/String;Ljava/lang/Class;)Ljava/lang/reflect/Method;");
-    DCHECK(getMethodObject);
-    ASSIGN_OR_RETURN(jstring method_name, helper.to_jstring(method.c_str()));
-    LOCAL_REF_GUARD(method_name);
-
-    jobject method_object = env->CallStaticObjectMethod(class_analyzer, getMethodObject, method_name, (jobject)clazz);
-    LOCAL_REF_GUARD(method_object);
-
-    RETURN_ERROR_IF_JNI_EXCEPTION_WITH_PREFIX(env, "call getMethodObject failed");
-
-    if (method_object == nullptr) {
-        return Status::InternalError(fmt::format("couldn't found method:{}", method));
-    }
-    auto res = env->NewGlobalRef(method_object);
-    return res;
-}
-
-Status ClassAnalyzer::get_method_desc(const std::string& sign, std::vector<MethodTypeDescriptor>* desc) {
-    RETURN_IF_ERROR(get_udaf_method_desc(sign, desc));
-    // return type may be a void type
-    for (int i = 1; i < desc->size(); ++i) {
-        if (desc->at(i).type == TYPE_UNKNOWN) {
-            return Status::InternalError(fmt::format("unknown type sign:{}", sign));
-        }
-    }
-    return Status::OK();
-}
-
-// clang-format off
-#define ADD_BOXED_METHOD_TYPE_DESC(STR, TYPE) \
-    } else if (type == STR) {                 \
-      desc->emplace_back(MethodTypeDescriptor{TYPE, true});
-
-#define ADD_PRIM_METHOD_TYPE_DESC(STR, TYPE) \
-    } else if (sign[i] == STR) {             \
-      desc->emplace_back(MethodTypeDescriptor{TYPE, false});
-// clang-format on
-
-Status ClassAnalyzer::get_udaf_method_desc(const std::string& sign, std::vector<MethodTypeDescriptor>* desc) {
-    for (int i = 0; i < sign.size(); ++i) {
-        if (sign[i] == '(' || sign[i] == ')') {
-            continue;
-        }
-        // Handle array types
-        if (sign[i] == '[') {
-            // Consume all leading '[' (for multi-dimensional arrays)
-            while (i < sign.size() && sign[i] == '[') {
-                ++i;
-            }
-
-            if (i >= sign.size()) {
-                return Status::InternalError(fmt::format("Invalid array descriptor '{}': missing element type", sign));
-            }
-
-            // Handle element type: object array 'L...;' or primitive type (single char)
-            // For varargs parameters (e.g. String..., Integer...), the method signature contains
-            // the array type ([Ljava/lang/String; etc.). Map the element type to the corresponding
-            // LogicalType so that get_method_desc() validation passes.
-            LogicalType elem_type = TYPE_UNKNOWN;
-            bool elem_is_box = true;
-            if (sign[i] == 'L') {
-                // Object array: [L<classname>;
-                ++i; // Skip 'L'
-                int elem_st = i;
-                while (i < sign.size() && sign[i] != ';') {
-                    ++i;
-                }
-                if (i >= sign.size()) {
-                    return Status::InternalError(
-                            fmt::format("Invalid object array descriptor '{}': missing ';'", sign));
-                }
-                std::string elem_class = sign.substr(elem_st, i - elem_st);
-                if (elem_class == "java/lang/String") {
-                    elem_type = TYPE_VARCHAR;
-                } else if (elem_class == "java/lang/Boolean") {
-                    elem_type = TYPE_BOOLEAN;
-                } else if (elem_class == "java/lang/Byte") {
-                    elem_type = TYPE_TINYINT;
-                } else if (elem_class == "java/lang/Short") {
-                    elem_type = TYPE_SMALLINT;
-                } else if (elem_class == "java/lang/Integer") {
-                    elem_type = TYPE_INT;
-                } else if (elem_class == "java/lang/Long") {
-                    elem_type = TYPE_BIGINT;
-                } else if (elem_class == "java/lang/Float") {
-                    elem_type = TYPE_FLOAT;
-                } else if (elem_class == "java/lang/Double") {
-                    elem_type = TYPE_DOUBLE;
-                } else if (elem_class == "java/util/List") {
-                    elem_type = TYPE_ARRAY;
-                } else if (elem_class == "java/util/Map") {
-                    elem_type = TYPE_MAP;
-                }
-                // i now points to ';', loop will increment it
-            } else {
-                // Primitive array: [Z [B [S [I [J [F [D
-                elem_is_box = false;
-                switch (sign[i]) {
-                case 'Z':
-                    elem_type = TYPE_BOOLEAN;
-                    break;
-                case 'B':
-                    elem_type = TYPE_TINYINT;
-                    break;
-                case 'S':
-                    elem_type = TYPE_SMALLINT;
-                    break;
-                case 'I':
-                    elem_type = TYPE_INT;
-                    break;
-                case 'J':
-                    elem_type = TYPE_BIGINT;
-                    break;
-                case 'F':
-                    elem_type = TYPE_FLOAT;
-                    break;
-                case 'D':
-                    elem_type = TYPE_DOUBLE;
-                    break;
-                default:
-                    break;
-                }
-                elem_type = TYPE_UNKNOWN;
-            }
-
-            desc->emplace_back(MethodTypeDescriptor{elem_type, elem_is_box, true});
-            continue;
-        }
-        if (sign[i] == 'L') {
-            int st = i + 1;
-            while (i < sign.size() && sign[i] != ';') {
-                i++;
-            }
-            if (i >= sign.size()) {
-                return Status::InternalError(fmt::format("Invalid object type descriptor '{}': missing ';'", sign));
-            }
-            std::string type = sign.substr(st, i - st);
-            if (false) {
-                // clang-format off
-            ADD_BOXED_METHOD_TYPE_DESC("java/lang/Boolean", TYPE_BOOLEAN)
-            ADD_BOXED_METHOD_TYPE_DESC("java/lang/Byte", TYPE_TINYINT)
-            ADD_BOXED_METHOD_TYPE_DESC("java/lang/Short", TYPE_SMALLINT)
-            ADD_BOXED_METHOD_TYPE_DESC("java/lang/Integer", TYPE_INT)
-            ADD_BOXED_METHOD_TYPE_DESC("java/lang/Long", TYPE_BIGINT)
-            ADD_BOXED_METHOD_TYPE_DESC("java/lang/Float", TYPE_FLOAT)
-            ADD_BOXED_METHOD_TYPE_DESC("java/lang/Double", TYPE_DOUBLE)
-                // clang-format on
-            } else if (type == "java/lang/String") {
-                desc->emplace_back(MethodTypeDescriptor{TYPE_VARCHAR, true});
-            } else if (type == "java/util/List") {
-                desc->emplace_back(MethodTypeDescriptor{TYPE_ARRAY, true});
-            } else if (type == "java/util/Map") {
-                desc->emplace_back(MethodTypeDescriptor{TYPE_MAP, true});
-            } else if (type == "java/math/BigDecimal") {
-                // Structural placeholder only: `.type` is not dispatched on at runtime for
-                // BigDecimal params. The actual DECIMAL precision/scale is resolved from
-                // ctx->get_arg_type() / ctx->get_return_type() at the call site.
-                desc->emplace_back(MethodTypeDescriptor{TYPE_DECIMAL128, true});
-            } else if (type == "java/time/LocalDate") {
-                desc->emplace_back(MethodTypeDescriptor{TYPE_DATE, true});
-            } else if (type == "java/time/LocalDateTime") {
-                desc->emplace_back(MethodTypeDescriptor{TYPE_DATETIME, true});
-            } else {
-                // Any other object class is a user-declared record bound to a STRUCT
-                // parameter/return. The FE analyzer has already validated at
-                // CREATE FUNCTION time that the class is a record matching the SQL
-                // STRUCT type, and that no unrecognized class appears in a non-STRUCT
-                // slot, so this branch is safe to treat as TYPE_STRUCT (no further
-                // class-name interpretation is needed here).
-                desc->emplace_back(MethodTypeDescriptor{TYPE_STRUCT, true});
-            }
-            continue;
-        }
-        if (false) {
-            // clang-format off
-        ADD_PRIM_METHOD_TYPE_DESC('Z', TYPE_BOOLEAN)
-        ADD_PRIM_METHOD_TYPE_DESC('B', TYPE_TINYINT)
-        ADD_PRIM_METHOD_TYPE_DESC('S', TYPE_SMALLINT)
-        ADD_PRIM_METHOD_TYPE_DESC('I', TYPE_INT)
-        ADD_PRIM_METHOD_TYPE_DESC('J', TYPE_BIGINT)
-        ADD_PRIM_METHOD_TYPE_DESC('F', TYPE_FLOAT)
-        ADD_PRIM_METHOD_TYPE_DESC('D', TYPE_DOUBLE)
-            // clang-format on
-        } else if (sign[i] == 'V') {
-            desc->emplace_back(MethodTypeDescriptor{TYPE_UNKNOWN, false});
-        } else {
-            desc->emplace_back(MethodTypeDescriptor{TYPE_UNKNOWN, false});
-        }
-    }
-
-    if (desc->size() > 1) {
-        desc->insert(desc->begin(), (*desc)[desc->size() - 1]);
-        desc->erase(desc->begin() + desc->size() - 1);
-    }
-    return Status::OK();
-}
-
-JavaUDFContext::~JavaUDFContext() = default;
-
-int UDAFFunction::create() {
-    auto [env, helper] = JVMFunctionHelper::getInstanceWithEnv();
-    jmethodID create = _ctx->ctx->create->get_method_id();
-    auto obj = env->CallObjectMethod(_udaf_handle, create);
-    LOCAL_REF_GUARD(obj);
-    CHECK_UDF_CALL_EXCEPTION(env, _function_context);
-    return _ctx->states->add_state(_function_context, env, obj);
-}
-
-void UDAFFunction::destroy(int state) {
-    auto [env, helper] = JVMFunctionHelper::getInstanceWithEnv();
-    auto obj = helper.convert_handle_to_jobject(_function_context, state);
-    LOCAL_REF_GUARD(obj);
-    jmethodID destory = _ctx->ctx->destory->get_method_id();
-    // call destroy
-    env->CallVoidMethod(_udaf_handle, destory, obj);
-    CHECK_UDF_CALL_EXCEPTION(env, _function_context);
-    _ctx->states->remove(_function_context, env, state);
-}
-
-jvalue UDAFFunction::finalize(int state) {
-    auto& helper = JVMFunctionHelper::getInstance();
-    JNIEnv* env = helper.getEnv();
-    auto obj = helper.convert_handle_to_jobject(_function_context, state);
-    LOCAL_REF_GUARD(obj);
-    jmethodID finalize = _ctx->ctx->finalize->get_method_id();
-    jvalue res;
-    res.l = env->CallObjectMethod(_udaf_handle, finalize, obj);
-    CHECK_UDF_CALL_EXCEPTION(env, _function_context);
-    return res;
-}
-
-void AggBatchCallStub::batch_update_single(int num_rows, jobject state, jobject* input, int cols) {
-    jvalue jni_inputs[3 + cols];
-    jni_inputs[0].i = num_rows;
-    jni_inputs[1].l = _caller;
-    jni_inputs[2].l = state;
-    for (int i = 0; i < cols; ++i) {
-        jni_inputs[3 + i].l = input[i];
-    }
-    auto* env = JVMFunctionHelper::getInstance().getEnv();
-    env->CallStaticVoidMethodA(_stub_clazz.clazz(), env->FromReflectedMethod(_stub_method.handle()), jni_inputs);
-    CHECK_UDF_CALL_EXCEPTION(env, _ctx);
-}
-
-StatusOr<jobject> BatchEvaluateStub::batch_evaluate(int num_rows, jobject* input, int cols) {
-    jvalue jni_inputs[2 + cols];
-    jni_inputs[0].i = num_rows;
-    jni_inputs[1].l = _caller;
-    for (int i = 0; i < cols; ++i) {
-        jni_inputs[2 + i].l = input[i];
-    }
-    auto* env = JVMFunctionHelper::getInstance().getEnv();
-    auto res = env->CallStaticObjectMethodA(_stub_clazz.clazz(), env->FromReflectedMethod(_stub_method.handle()),
-                                            jni_inputs);
-    RETURN_ERROR_IF_JNI_EXCEPTION(env);
-    return res;
-}
-
 Status JavaListStub::add(jobject element) {
     auto [env, helper] = JVMFunctionHelper::getInstanceWithEnv();
     const auto& meta = helper.list_meta();
@@ -1301,87 +758,6 @@ StatusOr<size_t> JavaListStub::size() {
     auto res = env->CallIntMethod(_list, meta.list_size);
     RETURN_ERROR_IF_JNI_EXCEPTION(env);
     return res;
-}
-
-void UDAFFunction::update(jvalue* val) {
-    auto [env, helper] = JVMFunctionHelper::getInstanceWithEnv();
-    jmethodID update = _ctx->ctx->update->get_method_id();
-    env->CallVoidMethodA(_udaf_handle, update, val);
-    CHECK_UDF_CALL_EXCEPTION(env, _function_context);
-}
-
-void UDAFFunction::merge(int state, jobject buffer) {
-    auto [env, helper] = JVMFunctionHelper::getInstanceWithEnv();
-    auto obj = helper.convert_handle_to_jobject(_function_context, state);
-    LOCAL_REF_GUARD(obj);
-    jmethodID merge = _ctx->ctx->merge->get_method_id();
-    env->CallVoidMethod(_udaf_handle, merge, obj, buffer);
-    CHECK_UDF_CALL_EXCEPTION(env, _function_context);
-}
-
-void UDAFFunction::serialize(int state, jobject buffer) {
-    auto [env, helper] = JVMFunctionHelper::getInstanceWithEnv();
-    auto obj = helper.convert_handle_to_jobject(_function_context, state);
-    LOCAL_REF_GUARD(obj);
-    jmethodID serialize = _ctx->ctx->serialize->get_method_id();
-    env->CallVoidMethod(_udaf_handle, serialize, obj, buffer);
-    CHECK_UDF_CALL_EXCEPTION(env, _function_context);
-}
-
-int UDAFFunction::serialize_size(int state) {
-    auto [env, helper] = JVMFunctionHelper::getInstanceWithEnv();
-    auto obj = helper.convert_handle_to_jobject(_function_context, state);
-    LOCAL_REF_GUARD(obj);
-    jmethodID serialize_size = _ctx->ctx->serialize_size->get_method_id();
-    int sz = env->CallIntMethod(obj, serialize_size);
-    CHECK_UDF_CALL_EXCEPTION(env, _function_context);
-    return sz;
-}
-
-void UDAFFunction::reset(int state) {
-    auto [env, helper] = JVMFunctionHelper::getInstanceWithEnv();
-    auto obj = helper.convert_handle_to_jobject(_function_context, state);
-    LOCAL_REF_GUARD(obj);
-    jmethodID reset = _ctx->ctx->reset->get_method_id();
-    env->CallVoidMethod(_udaf_handle, reset, obj);
-    CHECK_UDF_CALL_EXCEPTION(env, _function_context);
-}
-
-jobject UDAFFunction::window_update_batch(int state, int peer_group_start, int peer_group_end, int frame_start,
-                                          int frame_end, int col_sz, jobject* cols) {
-    auto [env, helper] = JVMFunctionHelper::getInstanceWithEnv();
-    auto obj = helper.convert_handle_to_jobject(_function_context, state);
-    LOCAL_REF_GUARD(obj);
-
-    jmethodID window_update = _ctx->ctx->window_update->get_method_id();
-    jvalue jvalues[5 + col_sz];
-    jvalues[0].l = obj;
-    jvalues[1].j = peer_group_start;
-    jvalues[2].j = peer_group_end;
-    jvalues[3].j = frame_start;
-    jvalues[4].j = frame_end;
-
-    for (int i = 0; i < col_sz; ++i) {
-        jvalues[i + 5].l = cols[i];
-    }
-
-    jobject res = env->CallObjectMethodA(_udaf_handle, window_update, jvalues);
-    CHECK_UDF_CALL_EXCEPTION(env, _function_context);
-    return res;
-}
-
-Status detect_java_runtime() {
-    const char* p = std::getenv("JAVA_HOME");
-    if (p == nullptr) {
-        return Status::RuntimeError("env 'JAVA_HOME' is not set");
-    }
-    auto st = call_hdfs_scan_function_in_pthread([]() {
-        if (getJNIEnv() == nullptr) {
-            return Status::RuntimeError("couldn't get JNIEnv, please check your java runtime");
-        }
-        return Status::OK();
-    });
-    return st->get_future().get();
 }
 
 } // namespace starrocks
