+++ be/src/udf/java/jvm_function_helper.h
@@ -13,17 +13,20 @@
 // limitations under the License.
 
 #pragma once
+
+#include <cstdint>
 #include <memory>
+#include <string>
 #include <unordered_map>
 #include <utility>
 
 #include "base/string/slice.h"
 #include "common/status.h"
 #include "common/statusor.h"
-#include "exprs/function_context.h"
 #include "jni.h"
 #include "runtime/env/java/java_env.h"
 #include "types/logical_type.h"
+#include "udf/java/jni_util.h"
 #include "udf/java/type_traits.h"
 
 #define DEFINE_JAVA_PRIM_TYPE(TYPE) \
@@ -37,9 +40,12 @@
     jclass TYPE##_class() { return _class_##PRIM_CLAZZ; }
 
 namespace starrocks {
-class DirectByteBuffer;
+
 class AggBatchCallStub;
 class BatchEvaluateStub;
+class Column;
+class DirectByteBuffer;
+class FunctionContext;
 
 struct ListMeta {
     // List class
@@ -320,53 +326,6 @@ private:
     JVMClass* _function_states_clazz = nullptr;
 };
 
-// local object reference guard.
-// The objects inside are automatically call DeleteLocalRef in the life object.
-#define LOCAL_REF_GUARD(lref)                                                \
-    DeferOp VARNAME_LINENUM(guard)([&lref]() {                               \
-        if (lref) {                                                          \
-            JVMFunctionHelper::getInstance().getEnv()->DeleteLocalRef(lref); \
-            lref = nullptr;                                                  \
-        }                                                                    \
-    })
-
-#define LOCAL_REF_GUARD_ENV(env, lref)              \
-    DeferOp VARNAME_LINENUM(guard)([&lref, env]() { \
-        if (lref) {                                 \
-            env->DeleteLocalRef(lref);              \
-            lref = nullptr;                         \
-        }                                           \
-    })
-
-// check JNI Exception and set error in FunctionContext
-#define CHECK_UDF_CALL_EXCEPTION(env, ctx)                                         \
-    if (auto e = env->ExceptionOccurred()) {                                       \
-        LOCAL_REF_GUARD(e);                                                        \
-        std::string msg = JVMFunctionHelper::getInstance().dumpExceptionString(e); \
-        LOG(WARNING) << "Exception: " << msg;                                      \
-        ctx->set_error(msg.c_str());                                               \
-        env->ExceptionClear();                                                     \
-    }
-
-#define RETURN_ERROR_IF_JNI_EXCEPTION_WITH_PREFIX(env, prefix)                     \
-    if (auto e = env->ExceptionOccurred()) {                                       \
-        LOCAL_REF_GUARD(e);                                                        \
-        std::string err = JVMFunctionHelper::getInstance().dumpExceptionString(e); \
-        env->ExceptionClear();                                                     \
-        return Status::InternalError(fmt::format("{}, {}", prefix, err));          \
-    }
-
-#define RETURN_ERROR_IF_JNI_EXCEPTION(env) RETURN_ERROR_IF_JNI_EXCEPTION_WITH_PREFIX(env, "JNI Exception")
-
-#define RETURN_IF_JNI_EXCEPTION(env, errmsg, ret)                                                                   \
-    if (jthrowable jthr = env->ExceptionOccurred()) {                                                               \
-        LOCAL_REF_GUARD(jthr);                                                                                      \
-        std::string msg = fmt::format("{},{}", errmsg, JVMFunctionHelper::getInstance().dumpExceptionString(jthr)); \
-        LOG(WARNING) << msg;                                                                                        \
-        env->ExceptionClear();                                                                                      \
-        return ret;                                                                                                 \
-    }
-
 // Used for UDAF serialization and deserialization,
 // providing a C++ memory space for Java to access.
 // DirectByteBuffer does not hold ownership of this memory space
@@ -410,42 +369,6 @@ private:
     int _capacity;
 };
 
-class AggBatchCallStub {
-public:
-    static inline const char* stub_clazz_name = "com.starrocks.udf.gen.CallStub";
-    static inline const char* batch_update_method_name = "batchCallV";
-
-    AggBatchCallStub(FunctionContext* ctx, jobject caller, JVMClass&& clazz, JavaGlobalRef&& method)
-            : _ctx(ctx), _caller(caller), _stub_clazz(std::move(clazz)), _stub_method(std::move(method)) {}
-
-    FunctionContext* ctx() { return _ctx; }
-
-    void batch_update_single(int num_rows, jobject state, jobject* input, int cols);
-
-private:
-    FunctionContext* _ctx;
-    // UDAF object handle, owned by JavaUDAFUniqueContext
-    jobject _caller;
-    JVMClass _stub_clazz;
-    JavaGlobalRef _stub_method;
-};
-
-class BatchEvaluateStub {
-public:
-    static inline const char* stub_clazz_name = "com.starrocks.udf.gen.CallStub";
-    static inline const char* batch_evaluate_method_name = "batchCallV";
-
-    BatchEvaluateStub(jobject caller, JVMClass&& clazz, JavaGlobalRef&& method)
-            : _caller(caller), _stub_clazz(std::move(clazz)), _stub_method(std::move(method)) {}
-
-    StatusOr<jobject> batch_evaluate(int num_rows, jobject* input, int cols);
-
-private:
-    jobject _caller;
-    JVMClass _stub_clazz;
-    JavaGlobalRef _stub_method;
-};
-
 class JavaListStub {
 public:
     JavaListStub(jobject list) : _list(list) {}
@@ -456,240 +379,5 @@ public:
 private:
     jobject _list;
 };
-
-// UDAF State Lists
-// mapping a java object as a int index
-// use get method to
-// TODO: implement a Java binder to avoid using this class
-class UDAFStateList {
-public:
-    static inline const char* clazz_name = "com.starrocks.udf.FunctionStates";
-    UDAFStateList(JavaGlobalRef&& handle, JavaGlobalRef&& get, JavaGlobalRef&& batch_get, JavaGlobalRef&& add,
-                  JavaGlobalRef&& remove, JavaGlobalRef&& clear);
-
-    jobject handle() { return _handle.handle(); }
-
-    // get state with index state
-    jobject get_state(FunctionContext* ctx, JNIEnv* env, int state);
-
-    // batch get states
-    jobject get_state(FunctionContext* ctx, JNIEnv* env, jobject state_ids);
-
-    // add a state to StateList
-    int add_state(FunctionContext* ctx, JNIEnv* env, jobject state);
-
-    // remove a state from StateList
-    void remove(FunctionContext* ctx, JNIEnv* env, int state);
-
-    // clear all state in StateList
-    void clear(FunctionContext* ctx, JNIEnv* env);
-
-private:
-    JavaGlobalRef _handle;
-    JavaGlobalRef _get_method;
-    JavaGlobalRef _batch_get_method;
-    JavaGlobalRef _add_method;
-    JavaGlobalRef _remove_method;
-    JavaGlobalRef _clear_method;
-    jmethodID _get_method_id;
-    jmethodID _batch_get_method_id;
-    jmethodID _add_method_id;
-    jmethodID _remove_method_id;
-    jmethodID _clear_method_id;
-};
-
-// For loading UDF Class
-// Not thread safe
-class ClassLoader {
-public:
-    static const inline int BATCH_SINGLE_UPDATE = 1;
-    static const inline int BATCH_EVALUATE = 2;
-    // Handle
-    ClassLoader(std::string path) : _path(std::move(path)) {}
-    ~ClassLoader();
-
-    ClassLoader& operator=(const ClassLoader& other) = delete;
-    ClassLoader(const ClassLoader&) = delete;
-    // get class
-    StatusOr<JVMClass> getClass(const std::string& className);
-    // get batch call stub
-    // numActualVarArgs: actual number of varargs input columns; only meaningful when the method uses varargs
-    StatusOr<JVMClass> genCallStub(const std::string& stubClassName, jclass clazz, jobject method, int type,
-                                   int numActualVarArgs = 0);
-
-    Status init();
-
-private:
-    std::string _path;
-    jmethodID _get_class = nullptr;
-    jmethodID _get_call_stub = nullptr;
-    JavaGlobalRef _handle = nullptr;
-    JavaGlobalRef _clazz = nullptr;
-};
-
-struct MethodTypeDescriptor {
-    LogicalType type;
-    bool is_box;
-    bool is_array = false;
-};
-
-struct JavaMethodDescriptor {
-    std::string signature; // function signature
-    std::string name;      // function name
-    std::vector<MethodTypeDescriptor> method_desc;
-    JavaGlobalRef method = nullptr;
-    // thread safe
-    jmethodID get_method_id() const;
-};
-
-// Used to get function signatures
-class ClassAnalyzer {
-public:
-    ClassAnalyzer() = default;
-    ~ClassAnalyzer() = default;
-
-    // Strip generic type parameters from a JNI method signature.
-    // e.g. "(Ljava/util/List<java/lang/String>;)V" -> "(Ljava/util/List;)V"
-    static void strip_jni_generic_types(std::string* sign);
-
-    Status has_method(jclass clazz, const std::string& method, bool* has);
-    Status get_signature(jclass clazz, const std::string& method, std::string* sign);
-    Status get_method_desc(const std::string& sign, std::vector<MethodTypeDescriptor>* desc);
-    StatusOr<jobject> get_method_object(jclass clazz, const std::string& method_name);
-    Status get_udaf_method_desc(const std::string& sign, std::vector<MethodTypeDescriptor>* desc);
-};
-
-struct JavaUDFContext {
-    JavaUDFContext() = default;
-    ~JavaUDFContext();
-
-    std::unique_ptr<ClassLoader> udf_classloader;
-    std::unique_ptr<ClassAnalyzer> analyzer;
-    std::unique_ptr<BatchEvaluateStub> call_stub;
-
-    JVMClass udf_class = nullptr;
-    JavaGlobalRef udf_handle = nullptr;
-
-    // Java Method
-    std::unique_ptr<JavaMethodDescriptor> prepare;
-    std::unique_ptr<JavaMethodDescriptor> evaluate;
-    std::unique_ptr<JavaMethodDescriptor> close;
-
-    // Per-argument and return-type UdfTypeDesc Java objects mirroring the SQL
-    // TypeDescriptor of each slot, with the formal Java record class captured at
-    // every STRUCT slot. Built once at UDF context construction (walking
-    // Method.getGenericParameterTypes() / getGenericReturnType() in lockstep with
-    // the SQL type tree). Passed verbatim to the unified UDFHelper.writeResult and
-    // input-boxing entry points so the recursion lives entirely on the Java side.
-    std::vector<JavaGlobalRef> evaluate_arg_type_descs;
-    JavaGlobalRef evaluate_return_type_desc = nullptr;
-};
-
-// Shareable, cacheable UDAF class-level context.
-// Contains class references and method descriptors — no per-aggregation state.
-// Cached via UserFunctionCache::load_cacheable_java_udf.
-struct JavaUDAFSharedContext {
-    std::unique_ptr<ClassLoader> udf_classloader;
-
-    JVMClass udaf_class = nullptr;
-    JVMClass udaf_state_class = nullptr;
-
-    std::unique_ptr<JavaMethodDescriptor> create;
-    std::unique_ptr<JavaMethodDescriptor> destory;
-    std::unique_ptr<JavaMethodDescriptor> update;
-    std::unique_ptr<JavaMethodDescriptor> merge;
-    std::unique_ptr<JavaMethodDescriptor> finalize;
-    std::unique_ptr<JavaMethodDescriptor> serialize;
-    std::unique_ptr<JavaMethodDescriptor> serialize_size;
-
-    // Window function methods (optional)
-    std::unique_ptr<JavaMethodDescriptor> reset;
-    std::unique_ptr<JavaMethodDescriptor> window_update;
-    std::unique_ptr<JavaMethodDescriptor> get_values;
-
-    // Generated stub class/method — used to create a per-aggregator AggBatchCallStub
-    JVMClass update_stub_clazz = nullptr;
-    JavaGlobalRef update_stub_method = nullptr;
-
-    // FunctionStates method objects — looked up once, cloned per aggregator instance
-    JavaGlobalRef states_get_method = nullptr;
-    JavaGlobalRef states_batch_get_method = nullptr;
-    JavaGlobalRef states_add_method = nullptr;
-    JavaGlobalRef states_remove_method = nullptr;
-    JavaGlobalRef states_clear_method = nullptr;
-
-    // Per-argument UdfTypeDesc Java objects mirroring the SQL TypeDescriptor of each
-    // UDAF arg (read by `update`). Slots whose type subtree contains no STRUCT are
-    // null-handle. Built once at shared-context construction by walking
-    // update.getGenericParameterTypes() (with state_offset=1 for the leading State arg).
-    std::vector<JavaGlobalRef> update_arg_type_descs;
-    // UdfTypeDesc for `finalize`'s return type when the subtree contains a STRUCT;
-    // null-handle otherwise. Used by the UDAF finalize / batch_finalize paths to
-    // route through the unified UDFHelper.writeResult drain.
-    JavaGlobalRef finalize_return_type_desc = nullptr;
-};
-
-// Per-aggregator UDAF context stored as FunctionContext::THREAD_LOCAL.
-// Holds a reference to the shared class-level JavaUDAFSharedContext plus
-// mutable per-aggregation state.
-struct JavaUDAFUniqueContext;
-
-class UDAFFunction {
-public:
-    UDAFFunction(jobject udaf_handle, FunctionContext* function_ctx, JavaUDAFUniqueContext* ctx)
-            : _udaf_handle(udaf_handle), _function_context(function_ctx), _ctx(ctx) {}
-    // create a new state for UDAF
-    int create();
-    // destroy state
-    void destroy(int state);
-    // UDAF Update Function
-    void update(jvalue* val);
-    // UDAF merge
-    void merge(int state, jobject buffer);
-    void serialize(int state, jobject buffer);
-    // UDAF State serialize_size
-    int serialize_size(int state);
-    // UDAF finalize
-    jvalue finalize(int state);
-
-    // WindowFunction reset
-    void reset(int state);
-
-    // WindowFunction updateBatch
-    jobject window_update_batch(int state, int peer_group_start, int peer_group_end, int frame_start, int frame_end,
-                                int col_sz, jobject* cols);
-
-private:
-    jobject _convert_to_jobject(int state);
-
-    // not owned udaf function handle
-    jobject _udaf_handle;
-    FunctionContext* _function_context;
-    JavaUDAFUniqueContext* _ctx;
-};
-
-struct JavaUDAFUniqueContext {
-    // Shared, possibly cached class-level context
-    std::shared_ptr<JavaUDAFSharedContext> ctx;
-
-    // Per-aggregator UDAF object instance and its batch-update stub
-    JavaGlobalRef handle = nullptr;
-    std::unique_ptr<AggBatchCallStub> update_batch_call_stub;
-
-    // Per-aggregator mutable state
-    std::unique_ptr<UDAFStateList> states;
-    std::unique_ptr<UDAFFunction> _func;
-    std::unique_ptr<DirectByteBuffer> buffer;
-    std::vector<uint8_t> buffer_data;
-};
-
-// Java UDAF lifecycle management based on FunctionContext::THREAD_LOCAL state.
-JavaUDAFUniqueContext* get_java_udaf_context(FunctionContext* ctx);
-void attach_java_udaf_context(FunctionContext* ctx, std::unique_ptr<JavaUDAFUniqueContext> udaf_ctx);
-void clear_java_udaf_states(FunctionContext* ctx);
-void destroy_java_udaf_context(FunctionContext* ctx);
-
-// Check whether java runtime can work
-Status detect_java_runtime();
 
 } // namespace starrocks
